requirements.txt
Django>=5.0,<6.0
django-unfold>=0.17.0
asgiref>=3.8.0
sqlparse>=0.5.0
tzdata>=2024.1
djangorestframework>=3.15